+++ lib/src/makeappx.dart
@@ -1,13 +1,12 @@
 import 'dart:io';
 import 'package:cli_util/cli_logging.dart' show Logger;
+import 'package:get_it/get_it.dart';
 import 'configuration.dart';
 
 /// Use the makeappx.exe tool to generate manifest file
 class MakeAppx {
-  final Configuration _config;
-  final Logger _logger;
-
-  MakeAppx(this._config, this._logger);
+  final Logger _logger = GetIt.I<Logger>();
+  final Configuration _config = GetIt.I<Configuration>();
 
   Future<void> pack() async {
     _logger.trace('packing');