+++ lib/src/windows_build.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 import 'package:cli_util/cli_logging.dart' show Logger;
+import 'package:get_it/get_it.dart';
 
 import 'configuration.dart';
 
@@ -8,10 +9,8 @@ const mainCppPath = 'windows/runner/main.cpp';
 
 /// Handles windows files build steps
 class WindowsBuild {
-  final Configuration _config;
-  final Logger _logger;
-
-  WindowsBuild(this._config, this._logger);
+  final Logger _logger = GetIt.I<Logger>();
+  final Configuration _config = GetIt.I<Configuration>();
 
   /// Run "flutter build windows" command
   Future<void> build() async {